requirements.txt
python-binance
numpy
pandas